refactor: add CopilotChat.WebApi.Auth.Specialization namespace
